Remove unnecessary test in Deploy-20533E0401DSC.ps1
In Deploy-20533E0401DSC.ps1, the scripts generates a name for the storageAccount ($saName) then checks for its availability. If it's not available, it starts looping until it finds an available name.
It is unnecessary to do the first If, as the loop is a Do...Until construction and will always be executed once.